% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preranks_gr.R
\name{preranks_gr}
\alias{preranks_gr}
\alias{get_prerank_gr}
\title{Pre-rank functions to assess multivariate calibration of gridded forecasts}
\usage{
get_prerank_gr(y, x, prerank, return_rank = TRUE, ...)
}
\arguments{
\item{y}{multivariate observation (numeric matrix with d1 columns and d2 rows).}

\item{x}{samples from multivariate forecast distribution (numeric array with dimension (d1, d2, M)).}

\item{prerank}{the pre-rank function to be used. This is either a string from
a list of possible options (see details below), or a function.}

\item{return_rank}{logical specifying whether the rank should be returned
(rather than the vector of pre-ranks).}

\item{...}{additional arguments to the pre-rank function.}
}
\value{
Rank of the pre-rank transformed observation among the forecast sample (if
\code{return_rank = T}), or a vector of pre-ranks corresponding to the observation
and sample members (if \code{return_rank = F}).
}
\description{
Calculate multivariate ranks of gridded forecasts and observations
}
\details{
The argument \code{prerank} specifies which pre-rank function to use. This can
either be a string corresponding to one of several in-built options, or it can
be a user-specified function.

The in-built pre-rank functions currently available are the multivariate rank
(\code{prerank = "multivariate_rank"}), the average rank (\code{"average_rank"}),
the band-depth rank (\code{"band_depth"}), the mean (\code{"mean"}), the variance
(\code{"variance"}), the energy score (\code{"energy_score"}), and the
fraction of threshold exceedances (\code{"fte_rank"}).
Pre-rank functions will later be added for the variogram, isotropy, and
minimum spanning tree. See references for details.

If \code{prerank} is a function, it should convert a vector of dimension d, to
a single numeric value. Checks are in place to ensure this is satisfied. The
\code{prerank} function could also take additional inputs, in which case these
inputs should be included as additional arguments in \code{get_prerank}.
See examples below.
}
\examples{
d1 <- 5
d2 <- 6
M <- 10

y <- matrix(rnorm(d1*d2), nrow = d1, ncol = d2)
x <- array(rnorm(d1*d2*M), c(d1, d2, M))

get_prerank_gr(y, x, prerank = "average_rank", return_rank = FALSE)
get_prerank(as.vector(y), matrix(x, nrow = d1*d2, ncol = M), prerank = "average_rank", return_rank = FALSE)

get_prerank_gr(y, x, prerank = "variance", return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variance")

get_prerank_gr(y, x, prerank = "variogram", h = c(0, 1), return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variogram", h = c(0, -1), return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variogram", h = c(0, 1), std = FALSE, return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variogram", h = rbind(c(0, 1), c(1, 0)))

get_prerank_gr(y, x, prerank = "isotropy", return_rank = FALSE)
get_prerank_gr(y, x, prerank = "isotropy")
get_prerank_gr(y, x, prerank = "isotropy", h = 2)
get_prerank_gr(y, x, prerank = "isotropy", h = c(1, 2))

}
\references{
Allen, S., Ziegel, J. and D. Ginsbourger (2023):
`Assessing the calibration of multivariate probabilistic forecasts'.
\emph{arXiv preprint}. arXiv:2307.05846.
\doi{10.48550/arXiv.2307.05846}
}
\author{
Sam Allen
}
