% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{variogram_func}
\alias{variogram_func}
\alias{vario}
\alias{vario_mat}
\title{Compute empirical variogram}
\usage{
vario(y, w_mat, p)

vario_mat(y, h, p)
}
\arguments{
\item{y}{vector or matrix of values from which to calculate the variogram}

\item{w_mat}{matrix of weights corresponding to the pairs of dimensions}

\item{p}{exponent used when calculating the variogram}

\item{h}{the spatial lag at which to calculate the empirical variogram}
}
\value{
The variogram
}
\description{
Compute empirical variogram
}
\details{
\code{vario()} takes a vector \code{y} of length $d$ and a matrix \code{w_mat}
of dimension $d \times d$. The elements of \code{w_mat} are weights, and must
therefore be non-negative.

This calculates the double sum from $i = 1, \dots, d$ and $j = 1, \dots, d$
of the weighted difference between the components of \code{y}.

\code{vario_mat()} takes a matrix \code{y} of dimension $p \times q$ and a vector
\code{h} of length 2. \code{h} represents the spatial lag at which the spatial
variogam will be calculated.

In both cases, \code{p} is a single positive number that represents the exponent
used when calculating the variogram. This is typically set equal to 2.
}
