% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preranks.R
\name{preranks}
\alias{preranks}
\alias{get_prerank}
\title{Pre-rank functions for multivariate calibration}
\usage{
get_prerank(y, dat, prerank, return_rank = TRUE, t = NULL)
}
\arguments{
\item{y}{multivariate observation (numeric vector of length d).}

\item{dat}{samples from multivariate forecast distribution (numeric matrix with d rows).}

\item{prerank}{the pre-rank function to be used. See details for a list of options.}

\item{return_rank}{logical specifying whether the rank should be returned
(rather than the vector of pre-ranks). Default is TRUE.}

\item{t}{threshold for the FTE pre-rank function (single numeric value).}
}
\value{
Rank of the pre-rank transformed observation among the forecast sample (if
\code{prerank = F}), or a vector of pre-ranks corresponding to the observation
and sample members (if \code{prerank = T}).
}
\description{
Calulate pre-ranks of multivariate forecasts and observations
}
\details{
The pre-rank functions currently included are the multivariate rank
(\code{prerank = "multivariate_rank}), the average rank (\code{"average_rank"}),
the band-depth rank (\code{"band_depth"}), the mean (\code{"mean"}), the variance
(\code{"variance"}), the energy score (\code{"energy_score"}), and the
fraction of threshold exceedances (\code{fte_rank}). See references for details.

Pre-rank functions will also be added for the variogram, isotropy, and
minimum spanning tree.
}
\examples{
d <- 5
M <- 10

y <- as.vector(mvtnorm::rmvnorm(1, rep(0, d)))
dat <- t(mvtnorm::rmvnorm(M, rep(0, d)))

get_prerank(y, dat, prerank = "average_rank", return_rank = F)
get_prerank(y, dat, prerank = "average_rank")

get_prerank(y, dat, prerank = "variance", return_rank = F)
get_prerank(y, dat, prerank = "variance")

get_prerank(y, dat, prerank = "FTE", t = 0.5, return_rank = F)
get_prerank(y, dat, prerank = "FTE", t = 0.5) # ties are resolved at random
get_prerank(y, dat, prerank = "FTE", t = 0.5)


### use sapply to apply to several forecast cases
n <- 1000
y <- t(mvtnorm::rmvnorm(n, rep(0, d)))

# calibrated forecasts
dat <- array(t(mvtnorm::rmvnorm(n*M, rep(0, d))), c(d, M, n))
mvranks <- sapply(1:n, function(i) get_prerank(y[, i], dat[, , i], prerank = "mean"))
barplot(table(mvranks)) # observation is equally likely to take each rank

# miscalibrated mean
dat <- array(t(mvtnorm::rmvnorm(n*M, rep(-0.5, d))), c(d, M, n))
mvranks <- sapply(1:n, function(i) get_prerank(y[, i], dat[, , i], prerank = "mean"))
barplot(table(mvranks))
# forecast's under-estimate the mean, so the observation often has a higher rank
# when evaluated using the mean pre-rank function

# miscalibrated variance
dat <- array(t(mvtnorm::rmvnorm(n*M, sigma = 1.5*diag(d))), c(d, M, n))
mvranks <- sapply(1:n, function(i) get_prerank(y[, i], dat[, , i], prerank = "variance"))
barplot(table(mvranks)) # observation is more likely to take a higher rank
# forecast's under-estimate the variance, so the observation often has a higher variance rank
# when evaluated using the variance pre-rank function

}
\references{
Gneiting, T., Stanberry, L. I., Grimit, E. P., Held, L. and N. A. Johnson (2008):
`Assessing probabilistic forecasts of multivariate quantities, with an application to ensemble predictions of surface winds'.
\emph{Test} 17, 211-235.
\doi{10.1007/s11749-008-0114-x}

Thorarinsdottir, T. L., Scheuerer, M. and C. Heinz (2016):
`Assessing the calibration of high-dimensional ensemble forecasts using rank histograms'.
\emph{Journal of computational and graphical statistics} 25, 105-122.
\doi{10.1080/10618600.2014.977447}

Allen, S., Ziegel, J. and D. Ginsbourger (2023):
`Assessing the calibration of multivariate probabilistic forecasts'.
\emph{arXiv preprint}.
}
\author{
Sam Allen
}
