% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preranks_gr.R
\name{preranks_gr}
\alias{preranks_gr}
\alias{get_prerank_gr}
\title{Pre-rank functions to assess multivariate calibration of gridded forecasts}
\usage{
get_prerank_gr(y, x, prerank, return_rank = TRUE, ...)
}
\arguments{
\item{y}{multivariate observation (numeric matrix with d1 columns and d2 rows).}

\item{x}{samples from multivariate forecast distribution (numeric array with dimension (d1, d2, M)).}

\item{prerank}{the pre-rank function to be used. This is either a string from
a list of possible options (see details below), or a function.}

\item{return_rank}{logical specifying whether the rank should be returned
(rather than the vector of pre-ranks).}

\item{...}{additional arguments to the pre-rank function.}
}
\value{
Rank of the pre-rank transformed observation among the forecast sample (if
\code{return_rank = T}), or a vector of pre-ranks corresponding to the observation
and sample members (if \code{return_rank = F}).
}
\description{
Calculate multivariate ranks of gridded forecasts and observations
}
\details{
The argument \code{prerank} specifies which pre-rank function to use. This can
either be a string corresponding to one of several in-built options, or it can
be a user-specified function.

The in-built pre-rank functions currently available are the multivariate rank
(\code{prerank = "multivariate_rank"}), the average rank (\code{"average_rank"}),
the band-depth rank (\code{"band_depth"}), the mean (\code{"mean"}), the variance
(\code{"variance"}), the energy score (\code{"energy_score"}), the
fraction of threshold exceedances (\code{"FTE"}), the variogram (\code{"variogram"}),
and the isotropy of the variogram (\code{"isotropy"}). See references for details.

If \code{prerank} is a function, it should convert a matrix to
a single numeric value. Checks are in place to ensure this is satisfied. The
\code{prerank} function could also take additional inputs, in which case these
inputs should be included as additional arguments in \code{get_prerank_gr()}.
See examples below.

The FTE pre-rank requires a threshold parameter \code{t}, which must be a single real value.
The isotropy pre-rank function requires that a single integer \code{h} is provided,
denoting the lag at which the variograms are compared.

The variogram pre-rank function requires either an additional argument \code{h}
that specifies the multivariate lag at which to calculate the variogram, or an argument
\code{w} which is an array of non-negative weights assigned to the different pairs of
dimensions. The lag \code{h} should be a vector containing two integers, or a matrix with two
columns, in which case the variogram is calculated for each row of the given matrix
and the sum of the variograms is returned.

Assuming \code{y} is a matrix of dimension (d1, d2), the weight matrix \code{w} must be
an array with dimensions (d1, d2, d1, d2). The element in the (i, j, k, l) entry contains
the weight between the grid point (i, j) and the grid point (k, l). If both \code{h}
and \code{w} are provided, then only \code{h} will be used.

The variogram and isotropy pre-rank functions additionally require an argument \code{p} specifying
the exponent to be used in the variogram. The variogram also takes an input \code{std}
specifying whether the variogram should be standardised by the variance across the dimensions.
The exponent \code{p} is a positive real number, whereas \code{std} is a logical.
The default is \code{p = 2} and \code{std = TRUE}.
}
\examples{
d1 <- 5
d2 <- 6
M <- 10

y <- matrix(rnorm(d1*d2), nrow = d1, ncol = d2)
x <- array(rnorm(d1*d2*M), c(d1, d2, M))

# create array with weights that are inversely proportional to the distance
w <- array(NA, c(d1, d2, d1, d2))
for (i in 1:d1) {
  for (j in 1:d2) {
    w[i, j, , ] <- outer(1:d1, 1:d2, FUN = function(k, l) 1 / sqrt((i - k)^2 + (j - l)^2))
  }
}
w[is.infinite(w)] <- 0

get_prerank_gr(y, x, prerank = "average_rank", return_rank = FALSE)
get_prerank(as.vector(y), matrix(x, nrow = d1*d2, ncol = M),
            prerank = "average_rank", return_rank = FALSE)

get_prerank_gr(y, x, prerank = "variance", return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variance")

get_prerank_gr(y, x, prerank = "variogram", h = c(0, 1), return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variogram", h = c(0, -1), return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variogram", h = c(0, 1), std = FALSE, return_rank = FALSE)
get_prerank_gr(y, x, prerank = "variogram", h = rbind(c(0, 1), c(1, 0)))
get_prerank_gr(y, x, prerank = "variogram", w = w)

get_prerank_gr(y, x, prerank = "isotropy", return_rank = FALSE)
get_prerank_gr(y, x, prerank = "isotropy")
get_prerank_gr(y, x, prerank = "isotropy", h = 2)
get_prerank_gr(y, x, prerank = "isotropy", h = c(1, 2))

}
\references{
Allen, S., Ziegel, J. and D. Ginsbourger (2023):
`Assessing the calibration of multivariate probabilistic forecasts'.
\emph{arXiv preprint}. arXiv:2307.05846.
\doi{10.48550/arXiv.2307.05846}
}
\seealso{
\link{preranks} for pre-rank functions for multivariate forecasts
}
\author{
Sam Allen
}
